+++ src/__experimental__/components/form-field/form-field.component.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import propTypes from "@styled-system/prop-types";
 import FormFieldStyle, { FieldLineStyle } from "./form-field.style";
 import Label from "../label";
-import FieldHelp from "../field-help";
+import FieldHelp from "../../../__internal__/field-help";
 import OptionsHelper from "../../../utils/helpers/options-helper";
 import tagComponent from "../../../utils/helpers/tags";
 import Logger from "../../../utils/logger/logger";
