+++ src/__internal__/field-help/field-help.d.ts
@@ -1,0 +1,14 @@
+import * as React from "react";
+
+export interface FieldHelpProps {
+  /** Child elements */
+  children?: React.ReactNode;
+  /** When true, label is placed in line an input */
+  labelInline?: boolean;
+  /** Width of a label in percentage. Works only when labelInline is true */
+  labelWidth?: number;
+}
+
+declare function FieldHelp(props: FieldHelpProps): JSX.Element;
+
+export default FieldHelp;
