+++ src/__internal__/field-help/field-help.style.js
@@ -1,6 +1,6 @@
 import styled, { css } from "styled-components";
 import PropTypes from "prop-types";
-import BaseTheme from "../../../style/themes/base";
+import BaseTheme from "../../style/themes/base";
 
 const FieldHelpStyle = styled.span`
   display: block;
