+++ src/__internal__/field-help/index.d.ts
@@ -1,0 +1,1 @@
+export { default } from "./field-help";
